requirements.txt
pandas
numpy
sqlalchemy
psycopg2-binary
python-dotenv
requests
google-cloud-bigquery
pandas-gbq
dotenv
pyarrow
